Oops, fixed inputs not being able to undo
It would return an array of extra undo entries for being able to undo 
the popping-out of a block, but when that doesn't happen, it returns an 
array with just undefined in it, which breaks things. It now just 
doesn't return an array, which will be falsy, and all works happy!